refactor: tests for new typings
