+++ src/tests/studyClass.test.ts
@@ -1,8 +1,7 @@
 import supertest from 'supertest';
 import { app } from '../app';
 import { loginAdmin } from './helpers';
-import { Class, EducationTypes } from '../types';
-import { PostStudyClass } from '../validator/studyClass.validator';
+import { ZStudyClass } from '../validator/studyClass.validator';
 
 const api = supertest(app);
 const studyClassRoute = '/api/studyClass/';
@@ -14,10 +13,10 @@ beforeAll(async () => {
   await api.get('/testAuth').set('Cookie', [sessionId]).expect(200);
 });
 
-export const dummyClass: PostStudyClass = {
+export const dummyClass: ZStudyClass = {
   classId: 'BSC123',
-  class: Class.First,
-  educationType: EducationTypes.Literature,
+  class: '1',
+  educationType: 'Literature',
 };
 
 describe('CRUD of StudyClass', () => {
