+++ src/validator/activeSubject.validator.ts
@@ -1,5 +1,5 @@
 import { z } from 'zod';
-import { StudyClass } from './studyClass.validator';
+import { ZStudyClass } from './studyClass.validator';
 import { ZSubject } from './subject.validator';
 import { ZTeacher } from './teacher.validator';
 
@@ -9,10 +9,10 @@ import { ZTeacher } from './teacher.validator';
 // which is an auto incremented value in the DB,
 // sequelize has to omit this value.
 export const ZActiveSubject = z.object({
-  serial: z.number().positive(),
+  serial: z.number().positive().optional(),
   subjectId: ZSubject.shape.subjectId,
-  classId: StudyClass.shape.classId,
+  classId: ZStudyClass.shape.classId,
   teacherId: ZTeacher.shape.userId,
-  subjectSchedule: z.string().max(6),
+  subjectSchedule: z.string().max(6).optional(),
 });
 export type ZActiveSubject = z.infer<typeof ZActiveSubject>;
