+++ src/validator/studyClass.validator.ts
@@ -1,9 +1,9 @@
 import { z } from 'zod';
 import { ZEducationType, ZStudyYear } from './general.validator';
 
-export const StudyClass = z.object({
+export const ZStudyClass = z.object({
   classId: z.string().max(6),
   class: ZStudyYear.optional(),
   educationType: ZEducationType.optional(),
 });
-export type StudyClass = z.infer<typeof StudyClass>;
+export type ZStudyClass = z.infer<typeof ZStudyClass>;
